+++ cmd/pdatagen/internal/metrics_structs.go
@@ -141,8 +141,8 @@ var intGauge = &messageValueStruct{
 }
 
 var doubleGauge = &messageValueStruct{
-	structName:     "DoubleGauge",
-	description:    "// DoubleGauge represents the type of a double scalar metric that always exports the \"current value\" for every data point.",
+	structName:     "Gauge",
+	description:    "// Gauge represents the type of a double scalar metric that always exports the \"current value\" for every data point.",
 	originFullName: "otlpmetrics.Gauge",
 	fields: []baseField{
 		&sliceField{
@@ -169,8 +169,8 @@ var intSum = &messageValueStruct{
 }
 
 var doubleSum = &messageValueStruct{
-	structName:     "DoubleSum",
-	description:    "// DoubleSum represents the type of a numeric double scalar metric that is calculated as a sum of all reported measurements over a time interval.",
+	structName:     "Sum",
+	description:    "// Sum represents the type of a numeric double scalar metric that is calculated as a sum of all reported measurements over a time interval.",
 	originFullName: "otlpmetrics.Sum",
 	fields: []baseField{
 		aggregationTemporalityField,