+++ model/pdata/metrics_test.go
@@ -49,7 +49,7 @@ func TestCopyData(t *testing.T) {
 			},
 		},
 		{
-			name: "DoubleGauge",
+			name: "Gauge",
 			src: &otlpmetrics.Metric{
 				Data: &otlpmetrics.Metric_Gauge{
 					Gauge: &otlpmetrics.Gauge{},
@@ -65,7 +65,7 @@ func TestCopyData(t *testing.T) {
 			},
 		},
 		{
-			name: "DoubleSum",
+			name: "Sum",
 			src: &otlpmetrics.Metric{
 				Data: &otlpmetrics.Metric_Sum{
 					Sum: &otlpmetrics.Sum{},
@@ -106,12 +106,12 @@ func TestDataType(t *testing.T) {
 	assert.Equal(t, MetricDataTypeNone, m.DataType())
 	m.SetDataType(MetricDataTypeIntGauge)
 	assert.Equal(t, MetricDataTypeIntGauge, m.DataType())
-	m.SetDataType(MetricDataTypeDoubleGauge)
-	assert.Equal(t, MetricDataTypeDoubleGauge, m.DataType())
+	m.SetDataType(MetricDataTypeGauge)
+	assert.Equal(t, MetricDataTypeGauge, m.DataType())
 	m.SetDataType(MetricDataTypeIntSum)
 	assert.Equal(t, MetricDataTypeIntSum, m.DataType())
-	m.SetDataType(MetricDataTypeDoubleSum)
-	assert.Equal(t, MetricDataTypeDoubleSum, m.DataType())
+	m.SetDataType(MetricDataTypeSum)
+	assert.Equal(t, MetricDataTypeSum, m.DataType())
 	m.SetDataType(MetricDataTypeIntHistogram)
 	assert.Equal(t, MetricDataTypeIntHistogram, m.DataType())
 	m.SetDataType(MetricDataTypeHistogram)
@@ -238,8 +238,8 @@ func TestMetricAndDataPointCount(t *testing.T) {
 		ilm.AppendEmpty()
 	}
 	assert.EqualValues(t, 0, md.DataPointCount())
-	ilm.At(1).SetDataType(MetricDataTypeDoubleGauge)
-	doubleGauge := ilm.At(1).DoubleGauge()
+	ilm.At(1).SetDataType(MetricDataTypeGauge)
+	doubleGauge := ilm.At(1).Gauge()
 	doubleGauge.DataPoints().AppendEmpty()
 	ilm.At(3).SetDataType(MetricDataTypeIntHistogram)
 	intHistogram := ilm.At(3).IntHistogram()
@@ -262,7 +262,7 @@ func TestDataPointCountWithNilDataPoints(t *testing.T) {
 	intGauge := ilm.Metrics().AppendEmpty()
 	intGauge.SetDataType(MetricDataTypeIntGauge)
 	doubleGauge := ilm.Metrics().AppendEmpty()
-	doubleGauge.SetDataType(MetricDataTypeDoubleGauge)
+	doubleGauge.SetDataType(MetricDataTypeGauge)
 	intHistogram := ilm.Metrics().AppendEmpty()
 	intHistogram.SetDataType(MetricDataTypeIntHistogram)
 	doubleHistogram := ilm.Metrics().AppendEmpty()
@@ -270,7 +270,7 @@ func TestDataPointCountWithNilDataPoints(t *testing.T) {
 	intSum := ilm.Metrics().AppendEmpty()
 	intSum.SetDataType(MetricDataTypeIntSum)
 	doubleSum := ilm.Metrics().AppendEmpty()
-	doubleSum.SetDataType(MetricDataTypeDoubleSum)
+	doubleSum.SetDataType(MetricDataTypeSum)
 	assert.EqualValues(t, 0, metrics.DataPointCount())
 }
 
@@ -282,7 +282,7 @@ func TestOtlpToInternalReadOnly(t *testing.T) {
 				InstrumentationLibraryMetrics: []*otlpmetrics.InstrumentationLibraryMetrics{
 					{
 						InstrumentationLibrary: generateTestProtoInstrumentationLibrary(),
-						Metrics:                []*otlpmetrics.Metric{generateTestProtoIntGaugeMetric(), generateTestProtoDoubleSumMetric(), generateTestProtoDoubleHistogramMetric()},
+						Metrics:                []*otlpmetrics.Metric{generateTestProtoIntGaugeMetric(), generateTestProtoSumMetric(), generateTestProtoDoubleHistogramMetric()},
 					},
 				},
 			},
@@ -322,8 +322,8 @@ func TestOtlpToInternalReadOnly(t *testing.T) {
 	assert.EqualValues(t, "my_metric_double", metricDouble.Name())
 	assert.EqualValues(t, "My metric", metricDouble.Description())
 	assert.EqualValues(t, "ms", metricDouble.Unit())
-	assert.EqualValues(t, MetricDataTypeDoubleSum, metricDouble.DataType())
-	dsd := metricDouble.DoubleSum()
+	assert.EqualValues(t, MetricDataTypeSum, metricDouble.DataType())
+	dsd := metricDouble.Sum()
 	assert.EqualValues(t, AggregationTemporalityCumulative, dsd.AggregationTemporality())
 	doubleDataPoints := dsd.DataPoints()
 	assert.EqualValues(t, 2, doubleDataPoints.Len())
@@ -370,7 +370,7 @@ func TestOtlpToFromInternalReadOnly(t *testing.T) {
 				InstrumentationLibraryMetrics: []*otlpmetrics.InstrumentationLibraryMetrics{
 					{
 						InstrumentationLibrary: generateTestProtoInstrumentationLibrary(),
-						Metrics:                []*otlpmetrics.Metric{generateTestProtoIntGaugeMetric(), generateTestProtoDoubleSumMetric(), generateTestProtoDoubleHistogramMetric()},
+						Metrics:                []*otlpmetrics.Metric{generateTestProtoIntGaugeMetric(), generateTestProtoSumMetric(), generateTestProtoDoubleHistogramMetric()},
 					},
 				},
 			},
@@ -384,7 +384,7 @@ func TestOtlpToFromInternalReadOnly(t *testing.T) {
 				InstrumentationLibraryMetrics: []*otlpmetrics.InstrumentationLibraryMetrics{
 					{
 						InstrumentationLibrary: generateTestProtoInstrumentationLibrary(),
-						Metrics:                []*otlpmetrics.Metric{generateTestProtoIntGaugeMetric(), generateTestProtoDoubleSumMetric(), generateTestProtoDoubleHistogramMetric()},
+						Metrics:                []*otlpmetrics.Metric{generateTestProtoIntGaugeMetric(), generateTestProtoSumMetric(), generateTestProtoDoubleHistogramMetric()},
 					},
 				},
 			},
@@ -473,7 +473,7 @@ func TestOtlpToFromInternalIntGaugeMutating(t *testing.T) {
 	}, internal.MetricsToOtlp(md.InternalRep()))
 }
 
-func TestOtlpToFromInternalDoubleSumMutating(t *testing.T) {
+func TestOtlpToFromInternalSumMutating(t *testing.T) {
 	newLabels := NewStringMap().InitFromMap(map[string]string{"k": "v"})
 
 	md := MetricsFromInternalRep(internal.MetricsFromOtlp(&otlpcollectormetrics.ExportMetricsServiceRequest{
@@ -483,7 +483,7 @@ func TestOtlpToFromInternalDoubleSumMutating(t *testing.T) {
 				InstrumentationLibraryMetrics: []*otlpmetrics.InstrumentationLibraryMetrics{
 					{
 						InstrumentationLibrary: generateTestProtoInstrumentationLibrary(),
-						Metrics:                []*otlpmetrics.Metric{generateTestProtoDoubleSumMetric()},
+						Metrics:                []*otlpmetrics.Metric{generateTestProtoSumMetric()},
 					},
 				},
 			},
@@ -499,7 +499,7 @@ func TestOtlpToFromInternalDoubleSumMutating(t *testing.T) {
 	metric.SetUnit("1")
 	assert.EqualValues(t, "1", metric.Unit())
 	// Mutate DataPoints
-	dsd := metric.DoubleSum()
+	dsd := metric.Sum()
 	assert.EqualValues(t, 2, dsd.DataPoints().Len())
 	metric.SetDataType(MetricDataTypeDoubleSum)
 	doubleDataPoints := metric.DoubleSum().DataPoints()
@@ -668,7 +668,7 @@ func BenchmarkOtlpToFromInternal_PassThrough(b *testing.B) {
 				InstrumentationLibraryMetrics: []*otlpmetrics.InstrumentationLibraryMetrics{
 					{
 						InstrumentationLibrary: generateTestProtoInstrumentationLibrary(),
-						Metrics:                []*otlpmetrics.Metric{generateTestProtoIntGaugeMetric(), generateTestProtoDoubleSumMetric(), generateTestProtoDoubleHistogramMetric()},
+						Metrics:                []*otlpmetrics.Metric{generateTestProtoIntGaugeMetric(), generateTestProtoSumMetric(), generateTestProtoDoubleHistogramMetric()},
 					},
 				},
 			},
@@ -711,7 +711,7 @@ func BenchmarkOtlpToFromInternal_IntGauge_MutateOneLabel(b *testing.B) {
 	}
 }
 
-func BenchmarkOtlpToFromInternal_DoubleSum_MutateOneLabel(b *testing.B) {
+func BenchmarkOtlpToFromInternal_Sum_MutateOneLabel(b *testing.B) {
 	req := &otlpcollectormetrics.ExportMetricsServiceRequest{
 		ResourceMetrics: []*otlpmetrics.ResourceMetrics{
 			{
@@ -719,7 +719,7 @@ func BenchmarkOtlpToFromInternal_DoubleSum_MutateOneLabel(b *testing.B) {
 				InstrumentationLibraryMetrics: []*otlpmetrics.InstrumentationLibraryMetrics{
 					{
 						InstrumentationLibrary: generateTestProtoInstrumentationLibrary(),
-						Metrics:                []*otlpmetrics.Metric{generateTestProtoDoubleSumMetric()},
+						Metrics:                []*otlpmetrics.Metric{generateTestProtoSumMetric()},
 					},
 				},
 			},
@@ -729,7 +729,7 @@ func BenchmarkOtlpToFromInternal_DoubleSum_MutateOneLabel(b *testing.B) {
 	b.ResetTimer()
 	for n := 0; n < b.N; n++ {
 		md := MetricsFromInternalRep(internal.MetricsFromOtlp(req))
-		md.ResourceMetrics().At(0).InstrumentationLibraryMetrics().At(0).Metrics().At(0).DoubleSum().DataPoints().At(0).LabelsMap().Upsert("key0", "value2")
+		md.ResourceMetrics().At(0).InstrumentationLibraryMetrics().At(0).Metrics().At(0).Sum().DataPoints().At(0).LabelsMap().Upsert("key0", "value2")
 		newReq := internal.MetricsToOtlp(md.InternalRep())
 		if len(req.ResourceMetrics) != len(newReq.ResourceMetrics) {
 			b.Fail()
@@ -816,7 +816,7 @@ func generateTestProtoIntGaugeMetric() *otlpmetrics.Metric {
 		},
 	}
 }
-func generateTestProtoDoubleSumMetric() *otlpmetrics.Metric {
+func generateTestProtoSumMetric() *otlpmetrics.Metric {
 	return &otlpmetrics.Metric{
 		Name:        "my_metric_double",
 		Description: "My metric",