+++ service/collector.go
@@ -56,6 +56,17 @@ const (
 	Closed
 )
 
+// (Internal note) Collector Lifecycle:
+// - New constructs a new Collector.
+// - Run starts the collector and calls (*Collector).execute.
+// - execute calls setupConfigurationComponents to handle configuration.
+//   If configuration parser fails, collector's config can be reloaded.
+//   Collector can be shutdown if parser gets a shutdown error.
+// - execute runs runAndWaitForShutdownEvent and waits for a shutdown event.
+//   SIGINT and SIGTERM, errors, and (*Collector).Shutdown can trigger the shutdown events.
+// - Upon shutdown, pipelines are notified, then pipelines and extensions are shut down.
+// - Users can call (*Collector).Shutdown anytime to shutdown the collector.
+
 // Collector represents a server providing the OpenTelemetry Collector service.
 type Collector struct {
 	info    component.BuildInfo