+++ translator/internaldata/oc_to_metrics.go
@@ -162,8 +162,8 @@ func descriptorTypeToMetrics(t ocmetrics.MetricDescriptor_Type, metric pdata.Met
 		metric.SetDataType(pdata.MetricDataTypeIntGauge)
 		return pdata.MetricDataTypeIntGauge
 	case ocmetrics.MetricDescriptor_GAUGE_DOUBLE:
-		metric.SetDataType(pdata.MetricDataTypeDoubleGauge)
-		return pdata.MetricDataTypeDoubleGauge
+		metric.SetDataType(pdata.MetricDataTypeGauge)
+		return pdata.MetricDataTypeGauge
 	case ocmetrics.MetricDescriptor_CUMULATIVE_INT64:
 		metric.SetDataType(pdata.MetricDataTypeIntSum)
 		sum := metric.IntSum()
@@ -171,11 +171,11 @@ func descriptorTypeToMetrics(t ocmetrics.MetricDescriptor_Type, metric pdata.Met
 		sum.SetAggregationTemporality(pdata.AggregationTemporalityCumulative)
 		return pdata.MetricDataTypeIntSum
 	case ocmetrics.MetricDescriptor_CUMULATIVE_DOUBLE:
-		metric.SetDataType(pdata.MetricDataTypeDoubleSum)
-		sum := metric.DoubleSum()
+		metric.SetDataType(pdata.MetricDataTypeSum)
+		sum := metric.Sum()
 		sum.SetIsMonotonic(true)
 		sum.SetAggregationTemporality(pdata.AggregationTemporalityCumulative)
-		return pdata.MetricDataTypeDoubleSum
+		return pdata.MetricDataTypeSum
 	case ocmetrics.MetricDescriptor_CUMULATIVE_DISTRIBUTION:
 		metric.SetDataType(pdata.MetricDataTypeHistogram)
 		histo := metric.Histogram()
@@ -194,12 +194,12 @@ func setDataPoints(ocMetric *ocmetrics.Metric, metric pdata.Metric) {
 	switch metric.DataType() {
 	case pdata.MetricDataTypeIntGauge:
 		fillIntDataPoint(ocMetric, metric.IntGauge().DataPoints())
-	case pdata.MetricDataTypeDoubleGauge:
-		fillDoubleDataPoint(ocMetric, metric.DoubleGauge().DataPoints())
+	case pdata.MetricDataTypeGauge:
+		fillDoubleDataPoint(ocMetric, metric.Gauge().DataPoints())
 	case pdata.MetricDataTypeIntSum:
 		fillIntDataPoint(ocMetric, metric.IntSum().DataPoints())
-	case pdata.MetricDataTypeDoubleSum:
-		fillDoubleDataPoint(ocMetric, metric.DoubleSum().DataPoints())
+	case pdata.MetricDataTypeSum:
+		fillDoubleDataPoint(ocMetric, metric.Sum().DataPoints())
 	case pdata.MetricDataTypeHistogram:
 		fillDoubleHistogramDataPoint(ocMetric, metric.Histogram().DataPoints())
 	case pdata.MetricDataTypeSummary: